requirements.txt
streamlit==1.31.0
sounddevice==0.4.6
soundfile==0.12.1
librosa==0.10.1
matplotlib==3.7.2
scipy==1.11.3
fastdtw==0.3.4
gtts==2.4.0
requests==2.31.0
SpeechRecognition==3.10.0
python-Levenshtein==0.21.1
torch==2.1.0
transformers==4.34.0
epitran==1.17
pandas==2.1.1
openai-whisper==20231117
sentencepiece==0.1.99
google-cloud-storage==2.13.0
pyaudio
numpy
